requirements.txt
flask
flask-cors
torch
torchvision
pillow
gunicorn
pymongo
bcrypt
